refactoring TestPrometheusMetricsSink - new class under integration
