+++ hadoop-hdds/framework/src/test/java/org/apache/hadoop/hdds/server/http/TestPrometheusMetricsSink.java
@@ -17,23 +17,8 @@
  */
 package org.apache.hadoop.hdds.server.http;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
-
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.OutputStreamWriter;
-import org.apache.commons.lang3.StringUtils;
-import org.apache.hadoop.metrics2.MetricsInfo;
-import org.apache.hadoop.metrics2.MetricsSource;
-import org.apache.hadoop.metrics2.MetricsSystem;
-import org.apache.hadoop.metrics2.MetricsTag;
-import org.apache.hadoop.metrics2.annotation.Metric;
-import org.apache.hadoop.metrics2.annotation.Metrics;
-import org.apache.hadoop.metrics2.lib.DefaultMetricsSystem;
-import org.apache.hadoop.metrics2.lib.MutableCounterLong;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 /**
@@ -41,166 +26,11 @@ import org.junit.jupiter.api.Test;
  */
 public class TestPrometheusMetricsSink {
 
-  private MetricsSystem metrics;
-  private PrometheusMetricsSink sink;
+  private static PrometheusMetricsSink sink;
 
-  private static final MetricsInfo PORT_INFO = new MetricsInfo() {
-    @Override
-    public String name() {
-      return "PORT";
-    }
-
-    @Override
-    public String description() {
-      return "port";
-    }
-  };
-
-  private static final MetricsInfo COUNTER_INFO = new MetricsInfo() {
-    @Override
-    public String name() {
-      return "COUNTER";
-    }
-
-    @Override
-    public String description() {
-      return "counter";
-    }
-  };
-
-  private static final int COUNTER_1 = 123;
-  private static final int COUNTER_2 = 234;
-
-  @BeforeEach
-  public void init() {
-    metrics = DefaultMetricsSystem.instance();
-
-    metrics.init("test");
+  @BeforeAll
+  public static void setUp() {
     sink = new PrometheusMetricsSink();
-    metrics.register("Prometheus", "Prometheus", sink);
-  }
-
-  @AfterEach
-  public void tearDown() {
-    metrics.stop();
-    metrics.shutdown();
-  }
-
-  @Test
-  public void testPublish() throws IOException {
-    //GIVEN
-    TestMetrics testMetrics = metrics
-        .register("TestMetrics", "Testing metrics", new TestMetrics());
-
-    testMetrics.numBucketCreateFails.incr();
-
-    //WHEN
-    String writtenMetrics = publishMetricsAndGetOutput();
-
-    //THEN
-    Assertions.assertTrue(
-        writtenMetrics.contains(
-            "test_metrics_num_bucket_create_fails{context=\"dfs\""),
-        "The expected metric line is missing from prometheus metrics output"
-    );
-
-    metrics.unregisterSource("TestMetrics");
-  }
-
-  @Test
-  public void testPublishWithSameName() throws IOException {
-    // GIVEN
-    metrics.register("FooBar", "fooBar", (MetricsSource) (collector, all) -> {
-      collector.addRecord("RpcMetrics").add(new MetricsTag(PORT_INFO, "1234"))
-          .addGauge(COUNTER_INFO, COUNTER_1).endRecord();
-
-      collector.addRecord("RpcMetrics").add(new MetricsTag(
-          PORT_INFO, "2345")).addGauge(COUNTER_INFO, COUNTER_2).endRecord();
-    });
-
-    // WHEN
-    String writtenMetrics = publishMetricsAndGetOutput();
-
-    // THEN
-    Assertions.assertTrue(
-        writtenMetrics.contains("rpc_metrics_counter{port=\"2345\""),
-        "The expected metric line is missing from prometheus metrics output");
-
-    Assertions.assertTrue(
-        writtenMetrics.contains("rpc_metrics_counter{port=\"1234\""),
-        "The expected metric line is missing from prometheus metrics output");
-
-    metrics.unregisterSource("FooBar");
-  }
-
-  @Test
-  public void testTypeWithSameNameButDifferentLabels() throws IOException {
-    // GIVEN
-    metrics.register("SameName", "sameName",
-        (MetricsSource) (collector, all) -> {
-          collector.addRecord("SameName").add(new MetricsTag(PORT_INFO, "1234"))
-              .addGauge(COUNTER_INFO, COUNTER_1).endRecord();
-          collector.addRecord("SameName").add(new MetricsTag(PORT_INFO, "2345"))
-              .addGauge(COUNTER_INFO, COUNTER_2).endRecord();
-        });
-
-    // WHEN
-    String writtenMetrics = publishMetricsAndGetOutput();
-
-    // THEN
-    Assertions.assertEquals(1, StringUtils.countMatches(writtenMetrics,
-        "# TYPE same_name_counter"));
-
-    // both metrics should be present
-    Assertions.assertTrue(
-        writtenMetrics.contains("same_name_counter{port=\"1234\""),
-        "The expected metric line is present in prometheus metrics output");
-    Assertions.assertTrue(
-        writtenMetrics.contains("same_name_counter{port=\"2345\""),
-        "The expected metric line is present in prometheus metrics output");
-
-    metrics.unregisterSource("SameName");
-  }
-
-  /**
-   * Make sure Prometheus metrics start fresh after each flush.
-   * Publish and flush the metrics and then check that
-   * the unregistered metric is not present.
-   */
-  @Test
-  public void testRemovingStaleMetricsOnFlush() throws IOException {
-    // GIVEN
-    metrics.register("StaleMetric", "staleMetric",
-        (MetricsSource) (collector, all) -> {
-          collector.addRecord("StaleMetric")
-              .add(new MetricsTag(PORT_INFO, "1234"))
-              .addGauge(COUNTER_INFO, COUNTER_1).endRecord();
-        });
-
-    metrics.register("SomeMetric", "someMetric",
-        (MetricsSource) (collector, all) -> {
-          collector.addRecord("SomeMetric")
-              .add(new MetricsTag(PORT_INFO, "4321"))
-              .addGauge(COUNTER_INFO, COUNTER_2).endRecord();
-        });
-
-    // unregister metric
-    metrics.unregisterSource("StaleMetric");
-
-    // WHEN
-    // publish and flush metrics
-    String writtenMetrics = publishMetricsAndGetOutput();
-
-    // THEN
-    // The first metric shouldn't be present
-    Assertions.assertFalse(
-        writtenMetrics.contains("stale_metric_counter{port=\"1234\""),
-        "The expected metric line is present in prometheus metrics output");
-    Assertions.assertTrue(
-        writtenMetrics.contains("some_metric_counter{port=\"4321\""),
-        "The expected metric line is present in prometheus metrics output");
-
-    metrics.unregisterSource("SomeMetric");
   }
 
   @Test
@@ -250,27 +80,4 @@ public class TestPrometheusMetricsSink {
         "jvm_metrics_gc_time_millis_g1_young_generation",
         sink.prometheusName(recordName, metricName));
   }
-
-  private String publishMetricsAndGetOutput() throws IOException {
-    metrics.publishMetricsNow();
-
-    ByteArrayOutputStream stream = new ByteArrayOutputStream();
-    OutputStreamWriter writer = new OutputStreamWriter(stream, UTF_8);
-
-    sink.writeMetrics(writer);
-    writer.flush();
-
-    return stream.toString(UTF_8.name());
-  }
-
-  /**
-   * Example metric pojo.
-   */
-  @Metrics(about = "Test Metrics", context = "dfs")
-  private static class TestMetrics {
-
-    @Metric
-    private MutableCounterLong numBucketCreateFails;
-  }
-
 }
